Restructure sources and improve the build system
This change adds VERSION file to the top level of
the project that is then providin the version to
all places - Makefile, setup.py and spec file

It also moves the plugins that are needed for
tests to the tests/ subdirectory and consolidates
the documentation to doc/ directory.

It then adds OSCHEDPROXY_PLUGINS environment
variable that can be used to redefine the plugin
directory location and uses it in the tests.

Change-Id: I372e83f94c944056b2e5212c75139d23f05ae527
Signed-off-by: Martin Sivak <msivak@redhat.com>